requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
grpcio==1.59.2
grpcio-tools==1.59.2
idna==3.4
protobuf==4.25.0
python-dotenv==1.0.0
requests==2.31.0
school-service-grpc @ git+https://github.com/alexseipopov/capybaras_school_service_grpc.git@main
urllib3==2.0.7
